CRUD user

+++ api/resources/user.py
@@ -1,59 +1,65 @@
 
 from flask_restful import Resource
-from flask import  jsonify, request
+from flask import request
 
+from api.schemas.user import UserSchema
 from models.user import User
 from extensions import db
 
 
 class UserList(Resource):
     def get(self):
-        users = User.query.all()
-        return jsonify(users)
+        try:
+            schema = UserSchema(many=True)
+            users = User.query.all()
+            return schema.dump(users)
+        except Exception as e:
+            return {"message": str(e)}, 400
     
     def post(self):
+        try:
+            schema = UserSchema()
+            user = schema.load(request.get_json(), instance=User(), session=db.session)
+            db.session.add(user)
+            db.session.commit()
 
-        data = request.get_json()
-
-        user = User(
-            username=data.get('username'),
-            email=data.get('email'),
-            age=data.get('age'),
-            first_name=data.get('first_name'),
-            last_name=data.get('last_name')
-        ) 
-
-        db.session.add(user)
-        db.session.commit()
-
-        return jsonify(msg="User created", user=user)
+            return {"message": "User created successfully", "user": schema.dump(user)}, 201
+        except Exception as e:
+            return {"message": str(e)}, 400
 
 class UserDetail(Resource):
     def get(self, user_id):
-        user = User.query.get_or_404(user_id)
-        return jsonify(user)
-    
+        try:
+            schema = UserSchema()
+            user = User.query.get(user_id)
+            if user:
+                return {"user" : schema.dump(user)}, 200
+            return {"message": "User not found"}, 404
+        except Exception as e:
+            return {"message": str(e)}, 400
+
     def put(self, user_id):
-        user = User.query.get_or_404(user_id)
-        data = request.get_json()
+       
+            schema = UserSchema(partial=True)
+            user = User.query.get(user_id)
+            if user:
+                user = schema.load(request.get_json(), instance=user)
+                db.session.add(user)
+                db.session.commit()
+                return {"message": "User updated successfully", "user": schema.dump(user)}, 200
+            return {"message": "User not found"}, 404
+        
 
-        user.username = data['username']
-        user.email = data['email']
-        user.age = data['age']
-        user.first_name = data['first_name']
-        user.last_name = data['last_name']
-
-        db.session.commit()
-
-        return jsonify(user)
-    
     def delete(self, user_id):
-        user = User.query.get_or_404(user_id)
-
-        db.session.delete(user)
-        db.session.commit()
-
-        return '', 204
+        try:
+            user = User.query.get(user_id)
+            if user:
+                db.session.delete(user)
+                db.session.commit()
+                return {"message": "User deleted successfully"}, 200
+            return {"message": "User not found"}, 404
+        except Exception as e:
+            return {"message": str(e)}, 400
